requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.3.1
matplotlib>=3.4.3
tqdm>=4.62.3
scikit-image>=0.18.2
pandas>=1.3.3
flask>=2.0.1
gunicorn>=20.1.0
python-dotenv>=0.19.0
docker>=6.1.2
